+++ scripts/reference.js
@@ -22,6 +22,6 @@ function get_docs_strings() {
     return docs;
 }
 
-await exec('(let-env lips.env.__parent__ (load "./dist/std.xcb"))').then(() => {
+await exec('(letenv lips.env.__parent__ (load "./dist/std.xcb"))').then(() => {
     console.log(JSON.stringify(get_docs_strings()));
 });